+++ pages/event/index.js
@@ -1,6 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import Comment from '../../components/Comment/Comment';
 import Image from 'next/image';
+import Modal from '../../components/Event/Modal';
+
+
 
 async function retrieveEvent(parentObjectId) {
   const response = await fetch('http://34.210.145.64:8000/events/retrieve-event/' + parentObjectId);
@@ -40,21 +43,12 @@ function Event() {
 
   const [eventData, setEventData] = useState();
   const [hostsData, setHostsData] = useState();
-  const [counter, setCounter] = useState(0);
   const [modal, setModal] = useState(false);
+
 
   const parentObjectId = '6423680881ba99669b08fa7d';
   const userId = "64221158e1bbeb5fc205ed21"
 
-  //increase counter
-  const increase = () => {
-    setCounter(count => count + 1);
-  };
-
-  //decrease counter
-  const decrease = () => {
-    setCounter(count => count - 1);
-  };
 
   useEffect(() => {
     retrieveEvent(parentObjectId).then(
@@ -180,32 +174,7 @@ function Event() {
           </span>
         </div>
       </div>
-
-      {/* Modal */}
-      <div className={"relative z-10 " + (modal ? "" : "hidden")} aria-labelledby="modal-title" role="dialog" aria-modal="true">
-        <div className="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity"></div>
-        <div className="fixed inset-0 z-10 overflow-y-auto">
-          <div className="flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0">
-            <div className="relative transform overflow-hidden rounded-lg bg-white text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg">
-              <div className="bg-white px-4 pt-5 pb-4 sm:p-6 sm:pb-4">
-                <div className="float-right cursor-pointer" onClick={() => setModal(false)}>X</div>
-                <h1 className="font-bold">Complete your RSVP</h1>
-                <div>Are you bringing anyone?</div>
-                <div className="flex flex-row justify-between border-[1px] border-black max-w-full h-24 px-5 items-center">
-                  <span>{counter}</span>
-                  <div className='flex flex-row gap-5'>
-                    <button onClick={decrease}>-</button>
-                    <button onClick={increase}>+</button>
-                  </div>
-                </div>
-                <div className="max-w-full mt-4">
-                  <button type="button" className="w-full py-2 border-[1px] border-blue-400 text-blue-400 rounded-lg" onClick={() => saveAttendee(userId, parentObjectId)}>Submit</button>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
+      <Modal modal={modal} setModal={setModal} saveAttendee={saveAttendee} userId={userId} parentObjectId={parentObjectId} />
     </div>
   );
 }
